file/requirements.txt
Flask==3.0.0
boto3==1.34.69
python-dotenv==1.0.1
Flask-Cors==4.0.0
gunicorn==22.0.0
Werkzeug==3.0.1
pytest==8.0.0
elasticsearch==8.11.1
python-json-logger==2.0.7